Fix programming appending null bytes to small files
The cause of this was that, on encryption, the program first reads
"header-size" bytes from the file to see if the file is already
encrypted. If it does not find enough bytes, the file is obviously not
encrypted so it goes through with the encryption by encrypting the bytes
it did read. In doing this, it would write out the entire "header" array
it had parsed bytes into, not just as many as it had read.